third_party/WebKit/LayoutTests/editing/inserting/before-after-input-element.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<div id="log"></div>
<script>
test(() => assert_selection(
    '<div contenteditable>|<input></div>',
    'insertText foo',
    '<div contenteditable>foo|<input></div>'),
    'insert text before INPUT');

test(() => assert_selection(
    '<div contenteditable><input>|</div>',
    'insertText foo',
    '<div contenteditable><input>foo|</div>'),
    'insert text after INPUT');
</script>
